// src/order/types/enums/order.enum.ts
import {PharmacyRejectReasons} from '../../../pharmacy/types/enums/pharmacy.enum';

export enum BucketStatusEnum {
  WAITING_USER_PREVIEW_CONFIRMATION = 'WAITING_USER_PREVIEW_CONFIRMATION',
  USER_PREVIEW_CONFIRMED = 'USER_PREVIEW_CONFIRMED',
  WAITING_FOR_PHARMACY_CONFIRMATION = 'WAITING_FOR_PHARMACY_CONFIRMATION',
  PHARMACY_CONFIRMED = 'PHARMACY_CONFIRMED',
  PHARMACY_REJECTED = 'PHARMACY_REJECTED',
  WAITING_FOR_USER_PAYMENT_CONFIRMATION = 'WAITING_FOR_USER_PAYMENT_CONFIRMATION',
  USER_PAYMENT_CONFIRMED = 'USER_PAYMENT_CONFIRMED',
  USER_PAYMENT_REJECTED = 'USER_PAYMENT_REJECTED',
  PHARMACY_PROCESSING = 'PHARMACY_PROCESSING',
  PHARMACY_PREPARED = 'PHARMACY_PREPARED',
  WAITING_FOR_RIDER_CONFIRMATION = 'WAITING_FOR_RIDER_CONFIRMATION',
  RIDER_CONFIRMED = 'RIDER_CONFIRMED',
  RIDER_NOT_FOUND = 'RIDER_NOT_FOUND',
  RIDER_REJECTED = 'RIDER_REJECTED',
  RIDER_ARRIVED_TO_PHARMACY = 'RIDER_ARRIVED_TO_PHARMACY',
  RIDER_COLLECTING_THE_ORDER_FROM_PHARMACY = 'RIDER_COLLECTING_THE_ORDER_FROM_PHARMACY',
  RIDER_ORDER_PICKED_FROM_PHARMACY = 'RIDER_ORDER_PICKED_FROM_PHARMACY',
  RIDER_ON_THE_WAY_TO_USER = 'RIDER_ON_THE_WAY_TO_USER',
  RIDER_ARRIVED_TO_USER = 'RIDER_ARRIVED_TO_USER',
  RIDER_WAITING_FOR_USER = 'RIDER_WAITING_FOR_USER',
  USER_CAME_TO_RIDER_TO_PICK_ORDER = 'USER_CAME_TO_RIDER_TO_PICK_ORDER',
  RIDER_ASK_FOR_PAYMENT = 'RIDER_ASK_FOR_PAYMENT',
  RIDER_RECEIVED_PAYMENT = 'RIDER_RECEIVED_PAYMENT',
  BUCKET_DELIVERED = 'BUCKET_DELIVERED',
  BUCKET_COMPLETE = 'BUCKET_COMPLETE',
  BUCKET_TIME_UP = 'BUCKET_TIME_UP',
  RIDER_CONFIRMATION_TIME_UP = 'RIDER_CONFIRMATION_TIME_UP',
  BUCKET_CANCELLED_BY_PHARMACY_WITH_REASON = 'BUCKET_CANCELLED_BY_PHARMACY_WITH_REASON',
  BUCKET_CANCELLED_BY_USER_WITH_REASON = 'BUCKET_CANCELLED_BY_USER_WITH_REASON',
  BUCKET_CANCELLED_BY_SYSTEM_WITH_REASON = 'BUCKET_CANCELLED_BY_SYSTEM_WITH_REASON',
}

export enum OrderPaymentEnum {
  CASH_ON_DELIVERY = 'CASH_ON_DELIVERY',
  CARD_PAYMENT = 'CARD_PAYMENT',
}

export enum OrderStatusEnum {
  RUNNING = 'RUNNING',
  COMPLETED = 'COMPLETED',
  CANCELLED = 'CANCELLED'
}

export enum RiderBucketEnum {
  TIME_UP = 'TIME_UP',
  PENDING = 'PENDING',
  ACCEPTED = 'ACCEPTED',
  PICKED = 'PICKED',
  REJECTED = 'REJECTED',
  COMPLETED = 'COMPLETED'
}

export enum BucketType {
  PRESCRIPTION = 'PRESCRIPTION',
  GENERAL = 'GENERAL'
}

export enum UserRemoveBucketEnum {
  BUCKET_IS_COSTLY = 'BUCKET_IS_COSTLY',
  BUCKET_NOT_REQUIRED = 'BUCKET_NOT_REQUIRED',
  OTHER_REASON = 'OTHER_REASON'
}

export enum SystemRemoveBucketEnum {
  NO_PHARMACY_FOUND_BY_SYSTEM = 'NO_PHARMACY_FOUND_BY_SYSTEM'
}

export type ProceedWithoutBucketEnum = UserRemoveBucketEnum | PharmacyRejectReasons | SystemRemoveBucketEnum